emmm

+++ src/main/java/bcy/service/util/TokenUtil.java
@@ -3,6 +3,8 @@ package bcy.service.util;
 import java.util.Calendar;
 import java.util.HashMap;
 
+import org.apache.commons.logging.Log;
+
 import com.auth0.jwt.JWT;
 import com.auth0.jwt.JWTVerifier;
 import com.auth0.jwt.algorithms.Algorithm;
@@ -12,7 +14,8 @@ import com.auth0.jwt.interfaces.DecodedJWT;
 public class TokenUtil {
     private static final String SECRET = "banciyuan123";
 
-    public static String generateToken(Long id, Integer level) {
+    public static String generateToken(Integer id, Integer level) {
+
         Calendar calendar = Calendar.getInstance();
         calendar.add(Calendar.HOUR, 48);
